ci: split regression shards and stabilize gui smoke harness
