+++ docs/testing/phase2-visual-validation.md
@@ -2,24 +2,25 @@
 
 ## Automated checks
 
-Run the full test suite:
+Run the current GUI shell regression group:
 
-```bash
-python -m pytest tests/ -v
+```powershell
+.\tools\ci\run_pytest_targets.ps1 -GroupNames @("pr-gui-shell")
 ```
 
 Run Phase 2-focused tests only:
 
-```bash
-python -m pytest tests/test_theme_manager.py -v
-python -m pytest tests/test_gui_layout.py -v
-python -m pytest tests/test_plot_toolbar.py -v
+```powershell
+$env:UV_CACHE_DIR = ".uv-cache"
+uv run pytest tests\test_theme_manager.py -q
+uv run pytest tests\test_gui_layout_core.py -q
+uv run pytest tests\test_plot_toolbar.py -q
 ```
 
 ## Manual verification
 
-1. Launch the app with `python main.py`.
-2. Confirm the main toolbar shows a theme selector with `light`, `dark`, and `colorblind`.
+1. Launch the app with `uv run python main.py`.
+2. Confirm the main toolbar shows a theme selector with `light` and `dark`.
 3. Switch themes and verify the visualization canvas redraws immediately.
 4. Open the Visualization tab and confirm the left parameter dock does not overlap the plot area.
 5. Change multiple controls quickly and verify the plot updates once after the debounce delay.
@@ -28,6 +29,7 @@ python -m pytest tests/test_plot_toolbar.py -v
 
 ## Accessibility check
 
-Use the Okabe-Ito colorblind mode and validate the palette with a simulator:
+Validate that both light and dark themes preserve readable contrast for axis labels,
+legends, and toolbar controls:
 
 - https://www.color-blindness.com/coblis-color-blindness-simulator/
