+++ pytest.ini
@@ -9,3 +9,4 @@ markers =
     gui: PySide6 desktop GUI tests and layout smoke checks.
     slow: Long-running tests that are not ideal for tight edit-run loops.
     integration: Multi-module integration tests that exercise shared config, pipeline, or GUI flows.
+    pr_smoke: Curated high-signal regression coverage for pull request gating.
