+++ tests/test_gui_config_integration.py
@@ -7,7 +7,7 @@ import pytest
 from core.app_config import default_pipeline_params, load_yaml_config
 from gui.main_window import MainWindow
 
-pytestmark = [pytest.mark.gui, pytest.mark.integration]
+pytestmark = [pytest.mark.gui, pytest.mark.integration, pytest.mark.pr_smoke]
 
 
 def test_main_window_uses_shared_default_pipeline_params(qapp) -> None:
