+++ tests/test_gui_preset_manager.py
@@ -12,7 +12,7 @@ import yaml
 from core.app_config import load_yaml_config
 from gui.main_window import MainWindow
 
-pytestmark = [pytest.mark.gui, pytest.mark.integration]
+pytestmark = [pytest.mark.gui, pytest.mark.integration, pytest.mark.pr_smoke]
 
 
 def _sample_matrix() -> pd.DataFrame:
